requirements.txt
gspread
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
oauth2client
line-bot-sdk
Flask
requests